requirements.txt
Flask
pydub
Flask-Cors
flask-ngrok
pyngrok
google-generativeai
